refactor : AiChatController·AiSongController를 AiController로 통합
